Some assets stuff (sass/img/js/config)
As we're now running on Materialize, app/bootstrap contain some useless/not up-to-date features.
Instead of requiring Twitter Bootstrap, we now require Materialize.

Some scrolling behaviour have been added, and with them some line of code.
Some Sass for the blinking button, and some Js to automatically scroll down
So the menu become top-fixed (A materialize feature).
But regarding to client action, we sometimes just want to go serious,
so regarding to the route, we are automatically scrolled to the content.

Junk is a file where i put what I've done with effort, but which are now useless. Just in case...