Switch from Twitter-Bootstrap to Materialize-css
The title announce a lot of changements...
Anything related to bootstrap has been switched to materialize, so it does
affect as much as css, as html with new components. For example, the nav bar, which have been
entirely reimagined. It come with some Javascript features, now added with a particular
Blade directive => main: @stack / child: @push. It seems to be the cleanest way to add some js
without polluting the webpack mixing process.

Some Ajax is here, and much is coming.